requirements.txt
bokeh
dockerspawner
ipympl
ipywidgets
jedi-language-server
jupyter_bokeh
jupyterhub==${JUPYTERHUB_VERSION}
jupyterhub-nativeauthenticator
jupyterlab
jupyterlab-drawio
jupyterlab_execute_time
jupyterlab-lsp
jupyterlab-miami-nights
jupyterlab_theme_minimalist
jupyterlab-theme-solarized-dark
jupyterlab_widgets
jupyter-resource-usage
koalas
octave_kernel
openpyxl
plotly
pyspark
pyrichlet
qiskit[visualization]
tabulate
yahoofinancials
yahooquery
